Test fixup for 32-bit systems

+++ test/weights.jl
@@ -341,7 +341,7 @@ try
                 elseif x < .9 || !resize
                     i = rand(eachindex(v))
                     x = if x < .41
-                        reinterpret(Float64, rand(0:reinterpret(UInt, floatmin(Float64))))
+                        reinterpret(Float64, rand(0:reinterpret(UInt64, floatmin(Float64))))
                     else
                         min(exp(rand((.1, 7, 300))*randn()), floatmax(Float64))
                     end
